+++ build.sh
@@ -348,7 +348,6 @@ SOURCES=(
   src/llvm/locals.weave
   src/llvm/strings.weave
   src/llvm/expr.weave
-  src/llvm/loop-phi.weave
   src/llvm/stmt.weave
   src/llvm/fn.weave
   src/llvm/module.weave